requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
et-xmlfile==1.1.0
fonttools==4.34.4
idna==3.3
imutils==0.5.4
kiwisolver==1.4.4
matplotlib==3.5.3
numpy==1.23.0
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
pandas==1.4.3
pdf2image==1.16.0
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
six==1.16.0
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
transforms==0.1
typing_extensions==4.3.0
urllib3==1.26.9
xlrd==1.2.0
